pyproject.toml
[project]
name = "coasti"
version = "0.1.3"
description = "Installer for Coasti, the Open-Source Business Intelligence Framework"
readme = "README.md"
authors = [
    { name = "F. Paul Spitzner", email = "paul.spitzner@linkfish.eu" }
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.10,<3.15"

dependencies = [
    "python-dotenv>=1.0.1",
    "typer>=0.15.2",
    "PyYAML>=6.0.2",
    "ruamel.yaml>=0.18.10",
    "xkcdpass>=1.20",
    "copier>=9.11.3",
]

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "docs" },
    { include-group = "typed" },
    "pre-commit>=4.5.1",
]
lint = ["ruff>=0.14.14"]
typed = ["mypy>=1.19.1", "types-PyYAML"]
docs = []
test = ["pytest>=9.0.2", "pytest-cov>=7.0.0"]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "coasti"
module-root = "src"
include = ["coasti/_bundles/*.bundle"]

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "ISC", # flake8-implicit-str-concat
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "PTH", # Use pathlib instead of os.path
    "W",   # pycodestyle
]

[tool.pytest.ini_options]
addopts = "--cov=coasti"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_decorators = true
allow_redefinition = true

# make the cli available
[project.scripts]
coasti = "coasti:cli.app"

[project.urls]
"Home Page" = "https://coasti.org/"
Repository = "https://github.com/linkFISH-Consulting/coasti_installer"
GitHub = "https://github.com/linkFISH-Consulting/coasti_installer"
